+++ CHANGELOG.md
@@ -35,6 +35,7 @@
 - opl-template attribute is now watched and can be changed dynamically
 - Add Veo-Labs brand as a button in the control bar leading to the Veo-Labs web site. Button can be removed using "opl-veo-labs-icon" attribute
 - Small images of indexes can now be expressed using an URL plus x and y coordinates. Useful if you want to use sprites for your indexes small images
+- Player attribute "opl-indexes" has been added to show / hide indexes
 
 ## BUG FIXES